assume already upcasted in appropriate functions
convertContact can reuse its own upcasted bodies instead or
re-upcasting them.

setupFrictionConstraint and setupContactConstraint are called
indirectly or directly by convertContact. As convertContact upcasts at
the beginning of it's call it should be safe to assume these values are
converted already. As well, convertContact only calls these functions
with the variable storing the already converted value.

transform access functions to be direct access if that is all they do.
- cp's (ManifoldPoint)
  - getPositionWorldOnA() -> positionWorldOnA
  - getPositionWorldOnB() -> positionWorldOnB